+++ aws_emr_launch/constructs/lambdas/emr_lambdas.py
@@ -78,6 +78,29 @@ class OverrideClusterConfigs(core.Construct):
         return self._lambda_function
 
 
+class UpdateClusterTags(core.Construct):
+    def __init__(self, scope: core.Construct, id: str) -> None:
+        super().__init__(scope, id)
+
+        code = aws_lambda.Code.from_asset(_lambda_path('emr_utilities'))
+        stack = core.Stack.of(scope)
+
+        self._lambda_function = stack.node.try_find_child('UpdateClusterTags')
+        if self._lambda_function is None:
+            self._lambda_function = aws_lambda.Function(
+                stack,
+                'UpdateClusterTags',
+                code=code,
+                handler='update_cluster_tags.handler',
+                runtime=aws_lambda.Runtime.PYTHON_3_7,
+                timeout=core.Duration.minutes(1)
+            )
+
+    @property
+    def lambda_function(self) -> aws_lambda.Function:
+        return self._lambda_function
+
+
 class ParseJsonString(core.Construct):
     def __init__(self, scope: core.Construct, id: str) -> None:
         super().__init__(scope, id)